Move HasUuid interface
