--- e2e/model.ts
@@ -1,3 +1,0 @@
-export interface HasUuid {
-    uuid: string;
-}
