+++ e2e/tests/node-publish.e2e-spec.ts
@@ -1,6 +1,5 @@
 import * as page from '../page-objects/app.po';
 import * as actions from '../page-objects/node-actions.po';
-import * as editor from '../page-objects/node-editor.po';
 import { i18n } from '../testUtil';
 import { nodes } from '../uuids';
